# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pet Adoption Manager API",
    "version": "0.2.0",
    "description": "Week 5: complete CRUD for shelters and pets.\n\n[Sign in with Google](/auth/google), then return here.\nUse GET /auth/me to check your identity and application role.\nNew identities must complete POST /api/users.\n\nPublic GET operations do not require sign-in.\nProtected operations use the browser session cookie.\nBefore writes, obtain GET /auth/csrf and enter its token\nthrough Authorize → csrfToken.\n\nPOST and PUT validate permitted fields.\nPUT updates supplied fields and preserves omitted fields.\nUnknown fields and empty updates return 400.\n\nRemaining user and application writes are planned for Week 6."
  },
  "servers": [
    {
      "url": "/",
      "description": "The current local or Render server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "pa.sid",
        "description": "Created by Google sign-in. The browser manages this HttpOnly cookie; do not paste a cookie value into Swagger."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Token returned by GET /auth/csrf."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "status": {
                "type": "integer",
                "example": 400
              },
              "message": {
                "type": "string",
                "example": "Invalid document ID."
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "example": "ageMonths"
                    },
                    "message": {
                      "type": "string",
                      "example": "Age must be nonnegative."
                    }
                  }
                }
              }
            },
            "required": [
              "status",
              "message"
            ]
          }
        },
        "required": []
      },
      "UserCreate": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 80,
            "example": "Julian Becerra",
            "minLength": 2
          }
        },
        "required": [],
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$",
            "example": "507f1f77bcf86cd799439011",
            "readOnly": true
          },
          "displayName": {
            "type": "string",
            "maxLength": 80,
            "example": "Julian Becerra",
            "minLength": 2
          },
          "role": {
            "type": "string",
            "enum": [
              "adopter",
              "shelterStaff",
              "admin"
            ],
            "example": "adopter"
          },
          "shelterId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$",
            "example": "507f1f77bcf86cd799439011"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": []
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 80,
            "example": "Julian Becerra",
            "minLength": 2
          },
          "role": {
            "type": "string",
            "enum": [
              "adopter",
              "shelterStaff",
              "admin"
            ],
            "example": "adopter"
          },
          "shelterId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$",
            "example": "507f1f77bcf86cd799439011"
          }
        },
        "required": [],
        "minProperties": 1,
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "listUser",
        "summary": "List users",
        "description": "Lists require an administrator. Individual reads allow the owner or an administrator. OAuth identifiers and email are omitted from these GET responses. Results are paginated and ordered by ID. Unsupported query parameters return 400.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Records visible to the current caller; data may be empty.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID, query, or request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Access denied or invalid CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "postUser",
        "summary": "Complete your OAuth profile",
        "description": "Creates the current verified OAuth identity’s profile. Optional displayName may be supplied. Identity, email, and adopter role come from the server. A second creation attempt returns 409.",
        "x-implementation-status": "implemented",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Profile created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID, query, or request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Access denied or invalid CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate profile or planned relationship conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getUser",
        "summary": "Get one user",
        "description": "Lists require an administrator. Individual reads allow the owner or an administrator. OAuth identifiers and email are omitted from these GET responses.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "507f1f77bcf86cd799439011"
            },
            "description": "MongoDB document identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "One visible record.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID, query, or request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Access denied or invalid CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record not found or outside caller scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "putUser",
        "summary": "[Week 6 planned] PUT users",
        "description": "Current Week 5 implementation returns 501 after session and CSRF checks. Planned behavior: Lists require an administrator. Individual reads allow the owner or an administrator. OAuth identifiers and email are omitted from these GET responses. PUT updates permitted fields. Application status changes must follow ownership and review rules. Deletes must preserve referenced records.",
        "x-implementation-status": "planned",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "507f1f77bcf86cd799439011"
            },
            "description": "MongoDB document identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "Planned Week 6 success response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID, query, or request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Access denied or invalid CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Planned Week 6 response when the record does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate profile or planned relationship conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "501": {
            "description": "Current Week 5 response: operation is planned for Week 6.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "deleteUser",
        "summary": "[Week 6 planned] DELETE users",
        "description": "Current Week 5 implementation returns 501 after session and CSRF checks. Planned behavior: Lists require an administrator. Individual reads allow the owner or an administrator. OAuth identifiers and email are omitted from these GET responses. PUT updates permitted fields. Application status changes must follow ownership and review rules. Deletes must preserve referenced records.",
        "x-implementation-status": "planned",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "507f1f77bcf86cd799439011"
            },
            "description": "MongoDB document identifier."
          }
        ],
        "responses": {
          "204": {
            "description": "Planned Week 6 deletion success."
          },
          "400": {
            "description": "Invalid ID, query, or request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Access denied or invalid CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Planned Week 6 response when the record does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate profile or planned relationship conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "501": {
            "description": "Current Week 5 response: operation is planned for Week 6.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/google": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Start Google sign-in",
        "description": "Open this route in a browser. Do not start the OAuth redirect flow through Try it out.",
        "security": [],
        "responses": {
          "302": {
            "description": "Redirect to Google with OAuth state."
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Receive the Google OAuth callback",
        "description": "Google invokes this route. Passport validates OAuth state and authenticates the Google identity.",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Login established; redirect to /api-docs."
          },
          "400": {
            "description": "Invalid ID, query, or request data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Access denied or invalid CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get the current identity or profile",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current account. New identities have needsOnboarding=true and no local _id or role.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "pattern": "^[a-fA-F0-9]{24}$",
                          "example": "507f1f77bcf86cd799439011",
                          "readOnly": true
                        },
                        "displayName": {
                          "type": "string",
                          "maxLength": 80,
                          "example": "Julian Becerra",
                          "minLength": 2
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "adopter",
                            "shelterStaff",
                            "admin"
                          ],
                          "example": "adopter"
                        },
                        "shelterId": {
                          "type": "string",
                          "pattern": "^[a-fA-F0-9]{24}$",
                          "example": "507f1f77bcf86cd799439011"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "readOnly": true
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "readOnly": true
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        }
                      },
                      "required": []
                    },
                    "needsOnboarding": {
                      "type": "boolean"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/csrf": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get a CSRF token",
        "description": "Copy csrfToken into Authorize → csrfToken before using a protected write operation. Get a new token after signing in again.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "CSRF token for this session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Csrf"
                }
              }
            }
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Sign out",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "Session destroyed and cookie cleared."
          },
          "401": {
            "description": "Google sign-in is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Access denied or invalid CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server or database failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api-docs": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Open Swagger UI",
        "security": [],
        "responses": {
          "200": {
            "description": "Swagger HTML interface.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "301": {
            "description": "Normalize the trailing slash."
          }
        }
      }
    },
...
  }
}